enable travis ci

+++ spec/spec_helper.rb
@@ -1,25 +1,29 @@
 require 'rspec'
 require 'ap'
 
-spec_dir = File.dirname(File.expand_path(__FILE__))
-REDIS_CMD = "redis-server #{spec_dir}/redis-test.conf"
+# No need to start redis when running in Travis
+unless ENV['CI']
 
-puts "Starting redis for testing at localhost:9736..."
-puts `cd #{spec_dir}; #{REDIS_CMD}`
+  spec_dir = File.dirname(File.expand_path(__FILE__))
+  REDIS_CMD = "redis-server #{spec_dir}/redis-test.conf"
+  
+  puts "Starting redis for testing at localhost:9736..."
+  puts `cd #{spec_dir}; #{REDIS_CMD}`
+  
+  require 'resque'
+  Resque.redis = 'localhost:9736'
+  require 'resque-concurrent-restriction'
+  
+  # Schedule the redis server for shutdown when tests are all finished.
+  at_exit do
+    pid = File.read("#{spec_dir}/redis.pid").to_i rescue nil
+    system ("kill -9 #{pid}") if pid.to_i != 0
+    File.delete("#{spec_dir}/redis.pid") rescue nil
+    File.delete("#{spec_dir}/redis-server.log") rescue nil
+    File.delete("#{spec_dir}/dump.rdb") rescue nil
+  end
 
-require 'resque'
-Resque.redis = 'localhost:9736'
-require 'resque-concurrent-restriction'
-
-# Schedule the redis server for shutdown when tests are all finished.
-at_exit do
-  pid = File.read("#{spec_dir}/redis.pid").to_i rescue nil
-  system ("kill -9 #{pid}") if pid.to_i != 0
-  File.delete("#{spec_dir}/redis.pid") rescue nil
-  File.delete("#{spec_dir}/redis-server.log") rescue nil
-  File.delete("#{spec_dir}/dump.rdb") rescue nil
 end
-
 
 ##
 # Helper to perform job classes
